requirements.txt
# Web Server & Logic
Flask
pyserial
numpy

# Computer Vision
opencv-python-headless

# Hardware & Adafruit Bridge
Adafruit-Blinka
adafruit-circuitpython-busdevice
adafruit-circuitpython-servokit
RPi.GPIO
lgpio